APCPCgodot/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
websockets==12.0
python-dotenv==1.0.1
discord.py==2.3.2
# psycopg2-binary==2.9.9
# asyncpg==0.29.0
